requirements.txt
ansible
ansible-lint
molecule[docker,lint]
pytest-testinfra
passlib
molecule-docker
yamllint
flake8
